complite Hierarchy

+++ src/Global/Global.h
@@ -36,7 +36,7 @@
 #define MAX_CSIZE 400
 
 namespace daf {
-    using Size = uint32_t;
+    using Size = uint64_t;
     using CliqueSize = uint16_t;
     static constexpr Size UNDEFINED = std::numeric_limits<Size>::max();
